backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pgvector==0.2.5
alembic==1.13.1
redis==5.0.4
rq==1.16.2
rq-scheduler==0.13.1
pydantic==2.7.1
pydantic-settings==2.3.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
cryptography==42.0.7
httpx==0.27.0
anthropic==0.28.0
sentence-transformers==3.0.1
langchain==0.2.1
langchain-community==0.2.1
langgraph==0.1.1
pypdf==4.2.0
python-docx==1.1.2
supabase==2.4.3
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.130.0
python-dotenv==1.0.1
structlog==24.1.0
tiktoken==0.7.0
